session

+++ backend_JPA/demo/src/main/java/com/example/demo/repository/DailyTaskRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.DailyTask;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DailyTaskRepository extends JpaRepository<DailyTask, Long> {
+public interface DailyTaskRepository extends JpaRepository<DailyTask, Integer> {
 }
