+++ backend_JPA/demo/src/main/java/com/example/demo/repository/TaskTagRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.TaskTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskTagRepository extends JpaRepository<TaskTag, Long> {
+public interface TaskTagRepository extends JpaRepository<TaskTag, Integer> {
 }
